refactor antd tabs and table
